nlp/CH5/practice/requirements.txt
fastapi==0.115.0
httpx==0.27.2
numpy==1.26.4
openai==1.51.0
pypdf==5.1.0
python-dotenv
qdrant-client
sentence-transformers==3.1.1
uvicorn==0.30.6
